+++ api/middleware/oauth2Model.js
@@ -3,11 +3,12 @@ import crypto from 'crypto';
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 
-export default function oauth2Model({ db, secret }) {
-  const { EmailAuthorization, OAuthAuthorization, OAuthClient } = db.models;
+export default function oauth2Model({ db, grant, secret }) {
+  const { EmailAuthorization, OAuthToken, OAuthAuthorization, OAuthClient } = db.models;
 
   return {
-    async generateAccessToken(client, user, scope) {
+    async generateToken(client, user, scope, expiresIn = 10800) {
+      // expires in 3 hours by default
       return jwt.sign(
         {
           scopes: scope,
@@ -17,17 +18,25 @@ export default function oauth2Model({ db, secret }) {
         {
           issuer: 'appsemble-api',
           subject: `${user.id}`,
-          expiresIn: 10800, // expires in 3 hours
+          expiresIn,
         },
       );
+    },
+
+    async generateAccessToken(client, user, scope) {
+      return this.generateToken(client, user, scope);
     },
 
     async generateRefreshToken() {
       return crypto.randomBytes(40).toString('hex');
     },
 
+    async generateAuthorizationCode(client, user, scope) {
+      return this.generateToken(client, user, scope);
+    },
+
     async getAccessToken(accessToken) {
-      const token = await OAuthAuthorization.findOne({ where: { token: accessToken } });
+      const token = await OAuthToken.findOne({ where: { token: accessToken } });
 
       if (!token) {
         return null;
@@ -48,7 +57,7 @@ export default function oauth2Model({ db, secret }) {
     },
 
     async getRefreshToken(refreshToken) {
-      const token = await OAuthAuthorization.findOne({ where: { refreshToken } });
+      const token = await OAuthToken.findOne({ where: { refreshToken } });
 
       if (!token) {
         return null;
@@ -67,20 +76,53 @@ export default function oauth2Model({ db, secret }) {
       }
     },
 
+    async getAuthorizationCode(authorizationCode) {
+      const token = await OAuthAuthorization.findOne(
+        { where: { token: authorizationCode } },
+        { raw: true },
+      );
+
+      if (!token) {
+        return null;
+      }
+
+      const expiresAt = new Date();
+      expiresAt.setTime(expiresAt.getTime() + 3 * 60 * 60 * 1000); // The duration of the generated JWT.
+
+      return {
+        code: authorizationCode,
+        expiresAt,
+        user: { id: token.UserId },
+        client: { id: 'appsemble-editor' }, // XXX: Figure out how to determine the client ID properly.
+        scope: 'apps:read apps:write',
+      };
+    },
+
     async getClient(clientId, clientSecret) {
       const clause = clientSecret ? { clientId, clientSecret } : { clientId };
       const client = await OAuthClient.findOne({ where: clause });
+      const config = grant
+        ? Object.values(grant.config).find(
+            entry => entry.key === clientId && entry.secret === clientSecret,
+          )
+        : undefined;
 
-      if (!client) {
+      if (!client && !config) {
         return false;
       }
 
-      return {
-        id: client.clientId,
-        secret: client.clientSecret,
-        redirect_uris: [client.redirectUri],
-        grants: ['password', 'refresh_token'],
-      };
+      return config
+        ? {
+            id: config.key,
+            secret: config.secret,
+            grants: ['authorization_code', 'refresh_token'],
+          }
+        : {
+            id: client.clientId,
+            secret: client.clientSecret,
+            redirect_uris: [client.redirectUri],
+            grants: ['password', 'refresh_token', 'authorization_code'],
+          };
     },
 
     async getUser(username, password) {
@@ -94,7 +136,7 @@ export default function oauth2Model({ db, secret }) {
     },
 
     async saveToken(token, client, user) {
-      await OAuthAuthorization.create({
+      await OAuthToken.create({
         token: token.accessToken,
         refreshToken: token.refreshToken,
         UserId: user.id,
@@ -107,13 +149,22 @@ export default function oauth2Model({ db, secret }) {
       };
     },
 
+    async saveAuthorizationCode(code, client, user) {
+      return this.saveToken(code, client, user);
+    },
+
     async revokeToken(token) {
       try {
-        await OAuthAuthorization.destroy({ where: { refreshToken: token.refreshToken } });
+        await OAuthToken.destroy({ where: { refreshToken: token.refreshToken } });
         return true;
       } catch (e) {
         return false;
       }
+    },
+
+    async revokeAuthorizationCode() {
+      // we want to manage these manually.
+      return true;
     },
 
     // XXX: Implement when implementing scopes