+++ api/models/OAuthAuthorization.js
@@ -2,9 +2,12 @@ export default (sequelize, DataTypes) =>
   sequelize.define(
     'OAuthAuthorization',
     {
-      id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
+      id: { type: DataTypes.STRING, primaryKey: true },
+      provider: { type: DataTypes.STRING, primaryKey: true },
       token: { type: DataTypes.TEXT, allowNull: false },
-      refreshToken: { type: DataTypes.TEXT, allowNull: false },
+      expiresAt: { type: DataTypes.DATE, allowNull: true },
+      refreshToken: { type: DataTypes.TEXT, allowNull: true },
+      verified: { type: DataTypes.BOOLEAN, default: false },
     },
     {
       freezeTableName: true,