+++ api/models/OAuthToken.js
@@ -1,0 +1,14 @@
+export default (sequelize, DataTypes) =>
+  sequelize.define(
+    'OAuthToken',
+    {
+      id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
+      token: { type: DataTypes.TEXT, allowNull: false },
+      refreshToken: { type: DataTypes.TEXT, allowNull: false },
+    },
+    {
+      freezeTableName: true,
+      createdAt: 'created',
+      updatedAt: 'updated',
+    },
+  );