+++ editor/components/App/App.jsx
@@ -1,17 +1,12 @@
 import { Loader } from '@appsemble/react-components';
-import { BrowserRouter as Router, Route, Switch, Link } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import React from 'react';
-import { IntlProvider, FormattedMessage } from 'react-intl';
+import { IntlProvider } from 'react-intl';
 
 import AppList from '../AppList';
 import Editor from '../Editor';
-import EmailLogin from '../EmailLogin';
-import ResetPassword from '../ResetPassword';
-import EditPassword from '../EditPassword';
+import Login from '../Login';
 import Message from '../Message';
-import Register from '../Register';
-import messages from './messages';
-import styles from './App.css';
 
 export default class App extends React.Component {
   async componentDidMount() {
@@ -22,7 +17,6 @@ export default class App extends React.Component {
   render() {
     const {
       user: { user, initialized },
-      authentication,
       logout,
     } = this.props;
 
@@ -35,32 +29,7 @@ export default class App extends React.Component {
         <div>
           <Router>
             {!user ? (
-              <Switch>
-                <Route component={Register} path="/editor/register" />
-                <Route component={ResetPassword} path="/editor/reset-password" />
-                <Route component={EditPassword} path="/editor/edit-password" />
-                <Route
-                  path="/editor"
-                  render={() => (
-                    <div>
-                      <EmailLogin
-                        authentication={{
-                          method: 'email',
-                          ...authentication,
-                        }}
-                      />
-                      <div className={styles.links}>
-                        <Link to="/editor/register">
-                          <FormattedMessage {...messages.registerLink} />
-                        </Link>
-                        <Link to="/editor/reset-password">
-                          <FormattedMessage {...messages.forgotPasswordLink} />
-                        </Link>
-                      </div>
-                    </div>
-                  )}
-                />
-              </Switch>
+              <Route render={props => <Login {...props} {...this.props} />} />
             ) : (
               <Switch>
                 <Route