+++ editor/components/Login/messages.jsx
@@ -1,0 +1,10 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  login: 'Login with {provider}',
+  register: 'Register',
+  greeting: 'Hello {name}!',
+  registerPrompt: 'Would you like to register an account using your {provider} credentials?',
+  registerLink: 'Register a new account',
+  forgotPasswordLink: 'Forgot your password?',
+});